# internal/model/entity/user_profile.go
package entity

type UserProfileEntity struct {
	Address string `json:"address"`
	Phone   string `json:"phone"`

	// fk
	UserID string `json:"user_id" gorm:"not null"`
}

type UserProfileModel struct {
	Entity
	UserProfileEntity
}

func (UserProfileModel) TableName() string {
	return "user_profiles"
}